Voice add function upload.

See merge request rapserver/smarthelper!507

+++ common-lib/src/main/java/com/rpa/common/mapper/VoiceShareMapper.java
@@ -8,5 +8,5 @@ import org.apache.ibatis.annotations.Mapper;
  * VoiceShareMapper继承基类
  */
 @Mapper
-public interface VoiceShareMapper extends BaseMapper<VoiceSharePO, Integer> {
+public interface VoiceShareMapper extends BaseMapper<VoiceSharePO, Long> {
 }
